date_separe

+++ ProjectCPP/GestionStock.h
@@ -10,10 +10,11 @@ public:
     void supprimer(void) override;
     void afficher(void) override;
     int changement_tarif(void);
+    //void ConvertDate(String^);
 
 protected:
 private:
-    int prixHT;
+    double prixHT;
     String^ reference;
     double TVA;
     int stock;
